无数据提示功能

+++ app/src/main/java/com/mt/androidtest_as/myrecyclerview/MyAdapter.java
@@ -21,12 +21,16 @@ import java.util.List;
  * Created by Mengtao1 on 2016/12/22.
  */
 
-public class BaseAdapter extends RecyclerView.Adapter<BaseViewHolder> implements View.OnClickListener{
+public class MyAdapter extends RecyclerView.Adapter<MyViewHolder> implements View.OnClickListener{
+    private static final int VIEW_TYPE_EMPTY = 0x01;
+    private static final int VIEW_TYPE_NOTEMPTY = 0x10;
     private Activity mActivity = null;
     private List<BaseData> mData = null;
-
-    public BaseAdapter(Activity mActivity){
+    private MyDataObserver mDataObserver = null;
+    public MyAdapter(Activity mActivity){
         this.mActivity = mActivity;
+        mDataObserver = new MyDataObserver();
+        registerAdapterDataObserver(mDataObserver);
     }
 
     public void setData(List<BaseData> mData){
@@ -34,24 +38,20 @@ public class BaseAdapter extends RecyclerView.Adapter<BaseViewHolder> implements
     }
 
     @Override
-    public BaseViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
-        if(null == mData||0 == mData.size()){
-            return null;
-        }
-        View mView = LayoutInflater.from(mActivity).inflate(R.layout.list_item, parent, false);
-        BaseViewHolder mBaseViewHolder = new BaseViewHolder(mView);
+    public MyViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        View mView = LayoutInflater.from(mActivity).inflate((viewType == VIEW_TYPE_EMPTY)?
+                R.layout.item_empty_view:R.layout.list_item, parent, false);
+        MyViewHolder mMyViewHolder = new MyViewHolder(mView, viewType);
+        mView.setTag(mMyViewHolder);
         mView.setOnClickListener(this);
-//        int position = mBaseViewHolder.getAdapterPosition();//position: Can noly get -1
-//        ALog.Log("position:"+position);
-        return mBaseViewHolder;
+        return mMyViewHolder;
     }
 
     @Override
-    public void onBindViewHolder(BaseViewHolder holder, int position) {
-        if(null == mData||0 == mData.size()){
+    public void onBindViewHolder(MyViewHolder holder, int position) {
+        if(VIEW_TYPE_EMPTY == holder.viewType){
             return;
         }
-        holder.mView.setTag(position);
         TextView mTextView = holder.mTextView;
         mTextView.setText(mData.get(position).getTitle());
     }
@@ -59,14 +59,35 @@ public class BaseAdapter extends RecyclerView.Adapter<BaseViewHolder> implements
     @Override
     public int getItemCount() {
         if(null == mData || 0 == mData.size()){
-            return 0;
+            return 1;
         }
         return mData.size();
     }
 
     @Override
+    public int getItemViewType(int position) {
+        //在这里进行判断，如果我们的集合的长度为0时，我们就使用emptyView的布局
+        if (null == mData || mData.size() == 0) {
+            return VIEW_TYPE_EMPTY;
+        }
+        //如果有数据，则使用ITEM的布局
+        return VIEW_TYPE_NOTEMPTY;
+    }
+
+    @Override
     public void onClick(View v) {
-        final int position = (int)v.getTag();
+        MyViewHolder holder = (MyViewHolder)v.getTag();
+        int viewType = holder.viewType;
+        if(VIEW_TYPE_EMPTY == viewType){
+            int initialDataNumber = 5;
+            DataBank.get(mActivity).generateData(initialDataNumber);
+            mData = DataBank.get(mActivity).getData();
+            setData(mData);
+            notifyDataSetChanged();
+            ALog.Log("onClick_Empty");
+            return;
+        }
+        final int position = holder.getAdapterPosition();
         Dialog mDialog = new AlertDialog.Builder(mActivity)
                 .setTitle(mActivity.getString(R.string.delete_item))
                 .setNegativeButton(android.R.string.cancel,null)
@@ -74,7 +95,7 @@ public class BaseAdapter extends RecyclerView.Adapter<BaseViewHolder> implements
                     @Override
                     public void onClick(DialogInterface dialog, int which) {
                         DataBank.get(mActivity).delCrime(mData.get(position));
-                        mData = DataBank.get(mActivity).getData();ALog.Log("size:"+mData.size());
+                        mData = DataBank.get(mActivity).getData();
                         setData(mData);
                         notifyDataSetChanged();
                     }
@@ -82,4 +103,5 @@ public class BaseAdapter extends RecyclerView.Adapter<BaseViewHolder> implements
                 .create();
         mDialog.show();
     }
+
 }
